fix: harden telegram html fallback text (#81764) (thanks @alexph-dev)
